Add docs input `Radio::class`.
