+++ docs/README.md
@@ -53,8 +53,9 @@ This is particularly useful for creating forms and other interactive elements on
 - [button-group](/docs/form/input/ButtonGroup.md)
 - [checkbox](/docs/form/input/Checkbox.md)
 - [date](/docs/form/input/Date.md)
-- [datetime-local](/docs/form/input/DatetimeLocal.md)
 - [datetime](/docs/form/input/Datetime.md)
+- [datetime-local](/docs/form/input/DatetimeLocal.md)
+- [radio](/docs/form/input/Radio.md)
 - [text](/docs/form/input/Text.md)
 - [time](/docs/form/input/Time.md)
 - [url](/docs/form/input/Url.md)
